# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Simple F\u0026B API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/bill/{id}": {
            "get": {
                "description": "get order by its id, returning the order details and the printers",
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves an order by its order id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BillResponse"
                        }
                    }
                }
            }
        },
        "/order": {
            "post": {
                "description": "create a new order by providing the table id and the details of the order, if success returns the newly created order id and a list of printers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "order request",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OrderResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.OrderDetailRequest": {
            "type": "object",
            "required": [
                "itemId",
                "itemType",
                "qty"
            ],
            "properties": {
                "itemId": {
                    "type": "integer"
                },
                "itemType": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
        "request.OrderRequest": {
            "type": "object",
            "required": [
                "details",
                "tableId"
            ],
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.OrderDetailRequest"
                    }
                },
                "tableId": {
                    "type": "integer"
                }
            }
        },
        "response.BillDetailResponse": {
            "type": "object",
            "properties": {
                "basePrice": {
                    "type": "number"
                },
                "itemName": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer"
                },
                "sumPrice": {
                    "type": "number"
                }
            }
        },
        "response.BillResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.BillDetailResponse"
                    }
                },
                "printers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Printer"
                    }
                },
                "table": {
                    "type": "string"
                },
                "totalPrice": {
                    "type": "number"
                }
            }
        },
        "response.OrderResponse": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "integer"
                },
                "printers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Printer"
                    }
                }
            }
        },
        "response.Printer": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}